Refactor index.ts into smaller files
